userspace counter improvements
Changes the internal behavior of the userspace counter wrapper that
is used to increment both a counter and histogram. This brings the
behavior more in-line with how the underlying Metriken counters
work and removes the potential footguns around previous value
tracking.
