+++ src/samplers/cpu/linux/usage/bpf.rs
@@ -224,10 +224,8 @@ impl CpuUsage {
 
         let now = Instant::now();
 
-        let mut total_busy = Counter::new(&CPU_USAGE_BUSY, Some(&CPU_USAGE_BUSY_HISTOGRAM));
-        let mut total_idle = Counter::new(&CPU_USAGE_IDLE, Some(&CPU_USAGE_IDLE_HISTOGRAM));
-        total_busy.set(0.0, 0);
-        total_idle.set(0.0, 0);
+        let total_busy = Counter::new(&CPU_USAGE_BUSY, Some(&CPU_USAGE_BUSY_HISTOGRAM));
+        let total_idle = Counter::new(&CPU_USAGE_IDLE, Some(&CPU_USAGE_IDLE_HISTOGRAM));
 
         Ok(Self {
             bpf,
@@ -251,16 +249,13 @@ impl CpuUsage {
 
         // update busy time metric
         let busy: u64 = busy();
-        let busy_prev = self.total_busy.value();
+        let busy_prev = self.total_busy.set(elapsed.as_secs_f64(), busy);
         let busy_delta = busy.wrapping_sub(busy_prev);
-        self.total_busy.set(elapsed.as_secs_f64(), busy);
 
         // calculate the idle time elapsed since last sample, update metric
-        let idle_prev = self.total_idle.value();
         let idle_delta =
             (self.online_cores.count as u64 * elapsed.as_nanos() as u64).saturating_sub(busy_delta);
-        let idle = idle_prev.wrapping_add(idle_delta);
-        self.total_idle.set(elapsed.as_secs_f64(), idle);
+        self.total_idle.add(elapsed.as_secs_f64(), idle_delta);
 
         // do the same for percpu counters
         for (cpu, (busy_counter, idle_counter)) in self
